requirements.txt
aiohttp==3.9.0
aiosignal==1.3.2
APScheduler==3.9.1
attrs==25.1.0
certifi==2024.12.14
charset-normalizer==2.1.1
click==8.1.8
Flask==2.2.5
Flask-SQLAlchemy==3.0.2
frozenlist==1.5.0
gunicorn==20.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.2.1
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
setuptools==75.8.0
six==1.17.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3
psycopg2-binary==2.9.6
flask-migrate==4.1.0
cryptography==41.0.7
Flask-WTF>=1.2.0
WTForms>=3.1.0
email-validator>=2.0.0
google-cloud-tasks>=2.14.0